requirements.txt
# Core dependencies
python-dotenv>=0.19.0
beautifulsoup4>=4.10.0
markdown>=3.3.6
pandas>=1.3.0
litellm>=0.8.1

# Error handling and retries
tenacity>=8.0.1

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
